完善3.1获取教师时间选择

+++ src/main/java/com/furp/service/TimeSlotsService.java
@@ -14,7 +14,7 @@ public interface TimeSlotsService {
 
     int updateDateConfigs(String year, List<TimeSlot> slots);
 
-    public List<TimeConfigVO> getTeacherAvailableTimeSlots(String year, Long teacherId);
+    public List<TimeConfigVO> getTeacherAvailableTimeSlots(String year, Integer teacherId);
 
     public List<AcademicTermVO> getAvailableAcademicTerms();
